Solve issue with shell's special characters (#7)
Solve the error that occurs when your settings contain one or more
characters that are considered as special characters by the shell. This
characters should be escaped or the command could fail or execute with
unexpected results.

This problem is described in issue #7

The proposed solution involves surrounding parameters that could contain
special characters with double quotes ("). This parameters / settings
are: user name, password and database name (you can create a user and a
database that contain special characters - on the other hand, I believe
it is not possible to have those special characters as a host name).

TODO: I haven't tested postgresql command because I don't have it
installed at this moment.

TODO 2: I don't have experience with oracle, so I didn't know how to
change the connection arguments or even if it was necessary.